+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPWKHJVNGWH8Y65XDVGJSKR"
+        "01KNR7R9DTX1DFRMP6NA4JSE69"
     )
     r"""The unique identifier of the data source"""