+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPWKHJSAMGCKPBCAKFSFCTT"
+        "01KNR7R9DQ32EAYHB3NY1SC2RB"
     )
     r"""The unique identifier of the data source"""