+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPWKHJTC4M3EEF7WVTCB5JN"
+        "01KNR7R9DS5VZND866EJZJEDQJ"
     )
     r"""The unique identifier of the data source"""