+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-08T15:50:15.158Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-09T04:24:20.915Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")