+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPWKHJY93Y4YJMSBJ550C92"
+        "01KNR7R9E2H580BWPG7NK9M5MJ"
     )
     r"""The unique identifier of the data source"""